implemented auth

+++ src/exercise/exercise.service.ts
@@ -4,6 +4,7 @@ import { IExercise, ExerciseTypeEnum } from './interfaces/exercise.interface';
 import { Model } from 'mongoose';
 import { CreateExerciseDto } from './dto/create-exercise.dto';
 import { UpdateExerciseDto } from './dto/update-exercise.dto';
+import { IUser, UserRoleEnum } from 'src/user/interfaces/user.interface';
 
 @Injectable()
 export class ExerciseService {
@@ -11,38 +12,86 @@ export class ExerciseService {
     @InjectModel('Exercise') private exerciseModel: Model<IExercise>,
   ) {}
 
-  async findAll(): Promise<IExercise[]> {
-    return this.exerciseModel.find();
+  async findAll(user: IUser): Promise<IExercise[]> {
+    return this.exerciseModel
+      .find({ $or: [{ owner: user._id }, { owner: null }] })
+      .sort('-createdAt');
   }
 
-  async findById(exerciseId: string): Promise<IExercise> {
-    return this.exerciseModel.findOne({ _id: exerciseId });
+  async findCompound(user: IUser): Promise<IExercise[]> {
+    return this.exerciseModel.find({
+      $and: [
+        { type: ExerciseTypeEnum.compound },
+        {
+          $or: [
+            {
+              owner: null,
+            },
+            {
+              owner: user._id,
+            },
+          ],
+        },
+      ],
+    });
   }
 
-  async findCompound(): Promise<IExercise[]> {
-    return this.exerciseModel.find({ type: ExerciseTypeEnum.compound });
+  async findIsolation(user: IUser): Promise<IExercise[]> {
+    return this.exerciseModel
+      .find({
+        $and: [
+          { type: ExerciseTypeEnum.isolation },
+          {
+            $or: [
+              {
+                owner: null,
+              },
+              {
+                owner: user._id,
+              },
+            ],
+          },
+        ],
+      })
+      .exec();
   }
 
-  async findIsolation(): Promise<IExercise[]> {
-    return this.exerciseModel.find({ type: ExerciseTypeEnum.isolation }).exec();
-  }
-
-  async create(createExerciseDto: CreateExerciseDto): Promise<IExercise> {
-    const newExercise = new this.exerciseModel(createExerciseDto);
+  async create(
+    createExerciseDto: CreateExerciseDto,
+    user: IUser,
+  ): Promise<IExercise> {
+    const newExercise = new this.exerciseModel({
+      ...createExerciseDto,
+      owner: user._id,
+    });
 
     return newExercise.save();
   }
 
-  async delete(exerciseId: string): Promise<IExercise> {
-    return this.exerciseModel.findByIdAndRemove(exerciseId);
+  async delete(exerciseId: string, user: IUser): Promise<IExercise> {
+    if (user.role === UserRoleEnum.admin) {
+      return this.exerciseModel.findByIdAndRemove(exerciseId);
+    }
+
+    return this.exerciseModel.findOneAndRemove({
+      $and: [{ owner: user._id }, { _id: exerciseId }],
+    });
   }
 
   async update(
     exerciseId: string,
     updateExerciseDto: UpdateExerciseDto,
+    user,
   ): Promise<IExercise> {
+    if (user.role === UserRoleEnum.admin) {
+      return this.exerciseModel.findByIdAndUpdate(
+        exerciseId,
+        updateExerciseDto,
+      );
+    }
+
     return this.exerciseModel.findOneAndUpdate(
-      { _id: exerciseId },
+      { $and: [{ owner: user._id }, { _id: exerciseId }] },
       updateExerciseDto,
       { new: true },
     );
